updating tests folder
