--- tests/ex0.c
@@ -1,3 +1,0 @@
-int main() {
-    return 0;
-}
