--- tests/ex3.c.out
@@ -1,37 +1,0 @@
-function foo(int m, int f)
-{
-KREG.1 = m + f;
-return KREG.1;
-}
-
-function main(int argc, char *)
-{
-int m;
-int a;
-float t;
-float f;
-int unreachable;
-
-goto <KREG.0>:;
-unreachable = 10;
-<KREG.0>::
-a = a + 1;
-a = a + 1;
-KREG.2 = foo2()
-KREG.3 = foo(KREG.2)
-KREG.4 = KREG.3 * 2;
-m = KREG.4;
-KREG.5 = m * a;
-a2 = KREG.5;
-KREG.6 = m * 5;
-KREG.7 = foo(KREG.6, f)
-KREG.8 = m * KREG.7;
-t = KREG.8;
-KREG.9 = m + 4;
-KREG.10 = 6 * 99;
-KREG.11 = KREG.9 + KREG.10;
-return KREG.11;
-}
-
-
-
