--- tests/ex4.c.out
@@ -1,37 +1,0 @@
-program 
-  funDeclaration 
-    typeSpecifier int
-    main
-    varDeclaration 
-      typeSpecifier int
-      varDeclList 
-        varDeclInitialize 
-          varDeclId j
-          =
-          constant 4
-    varDeclaration 
-      typeSpecifier float
-      varDeclList 
-        varDeclInitialize 
-          varDeclId k
-          =
-          constant 2
-    expression 
-      mutable f
-      =
-      relExpression 
-        constant 2
-        relop ==
-        constant 2
-    call printf
-      args 
-        argList 
-          argList 
-            constant "hello, world!"
-          ,
-          mutable f
-    returnStmt return
-      sumExpression 
-        call calcargs
-        sumop +
-        constant 2
