+++ tests/ex6.c
@@ -1,18 +1,40 @@
-#include <stdio.h>
+int main()
+{
+    char grade = 'B';
+    int num, curvedgrade;
 
-int main() {
-
-    int i, n, t1 = 0, t2 = 1, nextTerm;
-    printf("Enter the number of terms: ");
-    scanf("%d", &n);
-    printf("Fibonacci Series: ");
-
-    for (i = 1; i <= n; ++i) {
-        printf("%d, ", t1);
-        nextTerm = t1 + t2;
-        t1 = t2;
-        t2 = nextTerm;
+    switch(grade)
+    {
+        case 'A':
+            num = 1;
+            break;
+        case 'B':
+            num = 2;
+            break;
+        case 'C':
+            num = 3;
+            break;
+        case 'D':
+            num = 4;
+            break;
+        case 'F':
+            num = 5;
+            break;
+        default:
+            num = 6;
     }
+
+    if(grade == 'A') {
+        num = 90;
+    } else if(grade == 'B') {
+        num = 80;
+    } else if(grade == 'C') {
+        num = 70;
+    } else {
+        num = 60;
+    }
+
+    curvedgrade = num + 10;
+
     return 0;
-
-    }
+}
