+++ tests/ex7.c
@@ -1,47 +1,20 @@
-#include<stdio.h>
+int foo(int a, int b) {
+    int j = 100;
+    return j * (a+b);
+}
 
-int main()
-{
-    // Local Variable Definition
-    char grade;
-    int num, curvedgrade;
+int foobar() {
+    return 12 + 8 * 130;
+}
 
-    printf("Enter your grade:\n");
-    scanf("%c", &grade);
+int main(int argc) {
+    int i = 8, j = 2;
+    int k = 3;
+    int f = 100;
 
-    switch(grade)
-    {
-        case 'A':
-            printf("Excellent\n");
-            break;
-        case 'B':
-            printf("Keep it up!\n");
-            break;
-        case 'C':
-            printf("Well done\n");
-            break;
-        case 'D':
-            printf("You passed\n");
-            break;
-        case 'F':
-            printf("Better luck next time\n");
-            break;
-        default:
-            printf("Invalid grade\n");
-    }
-    printf("Your grade is %c\n",grade);
+    int a = i * (j + f * k);
+    int b = foobar();
+    int c = foo(a, b);
 
-    if(grade == 'A') {
-        num = 90;
-    } else if(grade == 'B') {
-        num = 80;
-    } else if(grade == 'C') {
-        num = 70;
-    } else {
-        num = 60;
-    }
-
-    curvedgrade = num + 10;
-
-    return 0;
+    return c * b + a;
 }
