+++ tests/ex8.c
@@ -1,20 +1,32 @@
-int foo(int a, int b) {
-    int j = 100;
-    return j * (a+b);
-}
+int main() {
 
-int foobar() {
-    return 12 + 8 * 130;
-}
+    int i;
+    int j = 0;
 
-int main(int argc) {
-    int i = 8, j = 2;
-    int k = 3;
-    int f = 100;
+    goto test;
+    i = 5;
 
-    int a = i * (j + f * k);
-    int b = foobar();
-    int c = foo(a, b);
+test:
+    while (i < 10) {
+        j *= 2;
+        i++;
+    }
 
-    return c * b + a;
+    if (j == 0) {
+        j = 0;
+    } else if (j == 9) {
+        j = 9;
+        i = 200;
+        if (i < 200) {
+            i = 0;
+        } else {
+            i = 500;
+        }
+    } else if (j == 8) {
+        j = 8;
+    } else {
+        j = 0;
+    }
+
+    return 0;
 }
