--- tests/ir_example2.ast
@@ -1,28 +1,0 @@
-program
-  - function main
-    - params
-    - body
-      - =
-        - j
-        - 0
-      - forloop
-        - =
-          - i
-          - 0
-        - <
-          - i
-          - 10
-        - ++
-          - i
-        - body
-          - if
-            - ==
-              - %
-                - i
-                - 2
-              - 0
-            - body
-              - ++
-                - j
-      - return
-        - j
